Correct implementation of stage0-module3-task5

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        int result = toBeDivided / 5;
+        double result = (double)toBeDivided / 5;
         System.out.println(result);
     }
 }
